# ARROW-11131: [Rust] Improve performance of bool_equal