This PR follows @jorgecarleitao 's great PR https://github.com/apache/arrow/pull/8541, with the following changes:

1. Implemented the logic when both `lhs` and `rhs` have zero null count.
2. May fixed a possible condition testing bug in `(0..len).all(|i| {...}` . 
    NOTE: there are other similar cases where `test_struct()` failed when updated in this way.
3. Added benchmarks `equal_bool_512` and `equal_bool_nulls_512`.

Benchmark comparing to arrow master:

```
equal_bool_512          time:   [52.946 ns 53.075 ns 53.188 ns]                           
                        change: [-96.295% -96.285% -96.277%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) low severe
  2 (2.00%) high mild
  2 (2.00%) high severe

equal_bool_nulls_512    time:   [2.3502 us 2.3696 us 2.3893 us]                                  
                        change: [-62.285% -62.019% -61.714%] (p = 0.00 < 0.05)
                        Performance has improved.
```